Website Correction + License (On ACO)
We are fixing the site and making it better, so sometimes there are such moments with mini fixes. We are just now fixing the site and the blockchain is almost ready, it remains to conduct tests, etc. And you can launch it, after its launch we will update the API: Cryptocurrency address, Blum link to the cryptocurrency and API price of cryptocurrency with DeDust through the pool, and convert the received price of TON to USD and we will display the price, change ±% and a graph with lines which will be convenient (+ when increasing you will see a green line and when decreasing a red one and the graph will be for 24 hours at 24 points per 1 hour)

ACO Licence - Anti-copying organization license (Anti-Copying organization) protection against theft and copying of the project, in case there will be a fine or even a criminal case